requirements.txt
# Core Dependencies
google-generativeai==0.3.2
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0

# Data Processing
pandas==2.1.1
numpy==1.24.3
PyPDF2==3.0.1
pdfplumber==0.9.0

# Web Framework
requests==2.31.0
werkzeug==2.3.7

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Additional utilities
python-json-logger==2.0.7
colorama==0.4.6
tqdm==4.66.1